Conflicts:
	sympy/core/function.py

+++ sympy/core/tests/test_evalf.py
@@ -2,6 +2,7 @@ from sympy import (Add, ceiling, cos, E, Eq, exp, factorial, fibonacci, floor,
                    Function, GoldenRatio, I, log, Mul, oo, pi, Pow, Rational,
                    sin, sqrt, sstr, Sum, sympify, S, integrate, atan, product)
 from sympy.core.evalf import complex_accuracy, PrecisionExhausted, scaled_zero
+from sympy.core.compatibility import long
 from sympy.mpmath import inf, ninf, nan
 from sympy.abc import n, x, y
 from sympy.mpmath.libmp.libmpf import from_float
@@ -234,9 +235,9 @@ def test_evalf_integer_parts():
     assert ceiling(10*(sin(1)**2 + cos(1)**2)) == 10
 
     assert int(floor(factorial(50)/E, evaluate=False).evalf(70)) == \
-        11188719610782480504630258070757734324011354208865721592720336800L
+        long(11188719610782480504630258070757734324011354208865721592720336800)
     assert int(ceiling(factorial(50)/E, evaluate=False).evalf(70)) == \
-        11188719610782480504630258070757734324011354208865721592720336801L
+        long(11188719610782480504630258070757734324011354208865721592720336801)
     assert int(floor((GoldenRatio**999 / sqrt(5) + Rational(1, 2)))
                .evalf(1000)) == fibonacci(999)
     assert int(floor((GoldenRatio**1000 / sqrt(5) + Rational(1, 2)))
@@ -401,8 +402,8 @@ def test_infinities():
 
 
 def test_to_mpmath():
-    assert sqrt(3)._to_mpmath(20)._mpf_ == (0, 908093L, -19, 20)
-    assert S(3.2)._to_mpmath(20)._mpf_ == (0, 838861L, -18, 20)
+    assert sqrt(3)._to_mpmath(20)._mpf_ == (0, long(908093), -19, 20)
+    assert S(3.2)._to_mpmath(20)._mpf_ == (0, long(838861), -18, 20)
 
 
 def test_issue_3533_evalf():
